requirements.txt
appdirs==1.4.4
astroid==2.8.2
awxkit==19.2.2
baron==0.9
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
debugpy==1.5.0
getgauge==0.3.17
ghp-import==2.0.1
grpcio==1.41.0
idna==3.2
importlib-metadata==4.6.4
isort==5.9.3
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mergedeep==1.3.4
mkdocs==1.2.2
mkdocs-material==7.2.4
mkdocs-material-extensions==1.0.1
packaging==21.0
platformdirs==2.4.0
protobuf==3.18.1
Pygments==2.10.0
pylint==2.11.1
pymdown-extensions==8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-owasp-zap-v2.4==0.0.18
PyYAML==5.4.1
pyyaml_env_tag==0.1
redbaron==0.9.2
requests==2.26.0
rply==0.7.8
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
watchdog==2.1.3
wrapt==1.12.1
zipp==3.5.0
